Moved EditionFilter to Admin namespace
