+++ test/unit/admin/edition_filter_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class EditionFilterTest < ActiveSupport::TestCase
+class Admin::EditionFilterTest < ActiveSupport::TestCase
   setup do
     @current_user = build(:gds_editor)
   end
@@ -9,14 +9,14 @@ class EditionFilterTest < ActiveSupport::TestCase
     policy = create(:policy)
     another_edition = create(:publication)
 
-    assert_equal [policy], EditionFilter.new(Edition, @current_user, type: 'policy').editions
+    assert_equal [policy], Admin::EditionFilter.new(Edition, @current_user, type: 'policy').editions
   end
 
   test "should filter by edition state" do
     draft_edition = create(:draft_policy)
     edition_in_other_state = create(:published_policy)
 
-    assert_equal [draft_edition], EditionFilter.new(Edition, @current_user, state: 'draft').editions
+    assert_equal [draft_edition], Admin::EditionFilter.new(Edition, @current_user, state: 'draft').editions
   end
 
   test "should filter by edition author" do
@@ -24,7 +24,7 @@ class EditionFilterTest < ActiveSupport::TestCase
     edition = create(:policy, authors: [author])
     edition_by_another_author = create(:policy)
 
-    assert_equal [edition], EditionFilter.new(Edition, @current_user, author: author.to_param).editions
+    assert_equal [edition], Admin::EditionFilter.new(Edition, @current_user, author: author.to_param).editions
   end
 
   test "should filter by organisation" do
@@ -33,7 +33,7 @@ class EditionFilterTest < ActiveSupport::TestCase
     edition_in_no_organisation = create(:policy)
     edition_in_another_organisation = create(:publication, organisations: [create(:organisation)])
 
-    assert_equal [edition], EditionFilter.new(Edition, @current_user, organisation: organisation.to_param).editions
+    assert_equal [edition], Admin::EditionFilter.new(Edition, @current_user, organisation: organisation.to_param).editions
   end
 
   test "should filter by edition type, state and author" do
@@ -41,7 +41,7 @@ class EditionFilterTest < ActiveSupport::TestCase
     policy = create(:draft_policy, authors: [author])
     another_edition = create(:published_policy, authors: [author])
 
-    assert_equal [policy], EditionFilter.new(Edition, @current_user, type: 'policy', state: 'draft', author: author.to_param).editions
+    assert_equal [policy], Admin::EditionFilter.new(Edition, @current_user, type: 'policy', state: 'draft', author: author.to_param).editions
   end
 
   test "should filter by edition type, state and organisation" do
@@ -49,7 +49,7 @@ class EditionFilterTest < ActiveSupport::TestCase
     policy = create(:draft_policy, organisations: [organisation])
     another_edition = create(:published_policy, organisations: [organisation])
 
-    assert_equal [policy], EditionFilter.new(Edition, @current_user, type: 'policy', state: 'draft', organisation: organisation.to_param).editions
+    assert_equal [policy], Admin::EditionFilter.new(Edition, @current_user, type: 'policy', state: 'draft', organisation: organisation.to_param).editions
   end
 
   test "should filter by edition type, state and world location" do
@@ -57,7 +57,7 @@ class EditionFilterTest < ActiveSupport::TestCase
     policy = create(:draft_policy, world_locations: [location])
     another_edition = create(:published_policy, world_locations: [location])
 
-    assert_equal [policy], EditionFilter.new(Edition, @current_user, type: 'policy', state: 'draft', world_location_ids: [location.id]).editions
+    assert_equal [policy], Admin::EditionFilter.new(Edition, @current_user, type: 'policy', state: 'draft', world_location_ids: [location.id]).editions
   end
 
   test "should filter by world location" do
@@ -65,103 +65,103 @@ class EditionFilterTest < ActiveSupport::TestCase
     consultation = create(:consultation)
     policy = create(:policy, world_locations: [location])
 
-    assert_equal [policy], EditionFilter.new(Edition, @current_user, world_location_ids: [location.id]).editions
+    assert_equal [policy], Admin::EditionFilter.new(Edition, @current_user, world_location_ids: [location.id]).editions
   end
 
   test "should filter by world location news article" do
     world_location_news_article = create(:world_location_news_article)
-    assert_equal [world_location_news_article], EditionFilter.new(Edition, @current_user, type: 'world_location_news_article').editions
+    assert_equal [world_location_news_article], Admin::EditionFilter.new(Edition, @current_user, type: 'world_location_news_article').editions
   end
 
   test "should filter by title" do
     detailed = create(:policy, title: "Test mcTest")
     policy = create(:policy, title: "A policy")
 
-    assert_equal [detailed], EditionFilter.new(Edition, @current_user, title: "test").editions
+    assert_equal [detailed], Admin::EditionFilter.new(Edition, @current_user, title: "test").editions
   end
 
   test "should return the editions ordered by most recent first" do
     older_policy = create(:draft_policy, updated_at: 3.days.ago)
     newer_policy = create(:draft_policy, updated_at: 1.minute.ago)
 
-    assert_equal [newer_policy, older_policy], EditionFilter.new(Edition, @current_user, {}).editions
+    assert_equal [newer_policy, older_policy], Admin::EditionFilter.new(Edition, @current_user, {}).editions
   end
 
   test "should be invalid if author can't be found" do
-    filter = EditionFilter.new(Edition, @current_user, author: 'invalid')
+    filter = Admin::EditionFilter.new(Edition, @current_user, author: 'invalid')
     refute filter.valid?
   end
 
   test "should be invalid if organisation can't be found" do
-    filter = EditionFilter.new(Edition, @current_user, organisation: 'invalid')
+    filter = Admin::EditionFilter.new(Edition, @current_user, organisation: 'invalid')
     refute filter.valid?
   end
 
   test "should generate page title when there are no filter options" do
-    filter = EditionFilter.new(Edition, build(:user))
+    filter = Admin::EditionFilter.new(Edition, build(:user))
     assert_equal "Everyone's documents", filter.page_title
   end
 
   test "should generate page title when we're displaying active documents" do
-    filter = EditionFilter.new(Edition, build(:user), state: 'active')
+    filter = Admin::EditionFilter.new(Edition, build(:user), state: 'active')
     assert_equal "Everyone's documents", filter.page_title
   end
 
   test "should generate page title when filtering by document state" do
-    filter = EditionFilter.new(Edition, build(:user), state: 'draft')
+    filter = Admin::EditionFilter.new(Edition, build(:user), state: 'draft')
     assert_equal "Everyone's draft documents", filter.page_title
   end
 
   test "should generate page title when filtering by document type" do
-    filter = EditionFilter.new(Edition, build(:user), type: 'news_article')
+    filter = Admin::EditionFilter.new(Edition, build(:user), type: 'news_article')
     assert_equal "Everyone's news articles", filter.page_title
   end
 
   test "should generate page title when filtering by any organisation" do
     organisation = create(:organisation, name: "Cabinet Office")
-    filter = EditionFilter.new(Edition, build(:user), organisation: organisation.to_param)
+    filter = Admin::EditionFilter.new(Edition, build(:user), organisation: organisation.to_param)
     assert_equal "Cabinet Office's documents", filter.page_title
   end
 
   test "should generate page title when filtering by my organisation" do
     organisation = create(:organisation)
     user = create(:user, organisation: organisation)
-    filter = EditionFilter.new(Edition, user, organisation: organisation.to_param)
+    filter = Admin::EditionFilter.new(Edition, user, organisation: organisation.to_param)
     assert_equal "My department's documents", filter.page_title
   end
 
   test "should generate page title when filtering by any author" do
     user = create(:user, name: 'John Doe')
-    filter = EditionFilter.new(Edition, build(:user), author: user.to_param)
+    filter = Admin::EditionFilter.new(Edition, build(:user), author: user.to_param)
     assert_equal "John Doe's documents", filter.page_title
   end
 
   test "should generate page title when filtering by my documents" do
     user = create(:user)
-    filter = EditionFilter.new(Edition, user, author: user.to_param)
+    filter = Admin::EditionFilter.new(Edition, user, author: user.to_param)
     assert_equal "My documents", filter.page_title
   end
 
   test "should generate page title when filtering by document state, document type and organisation" do
     organisation = create(:organisation, name: 'Cabinet Office')
-    filter = EditionFilter.new(Edition, build(:user), state: 'published', type: 'consultation', organisation: organisation.to_param)
+    filter = Admin::EditionFilter.new(Edition, build(:user), state: 'published', type: 'consultation', organisation: organisation.to_param)
     assert_equal "Cabinet Office's published consultations", filter.page_title
   end
 
   test "should generate page title when filtering by document state, document type and author" do
     user = create(:user, name: 'John Doe')
-    filter = EditionFilter.new(Edition, build(:user), state: 'rejected', type: 'speech', author: user.to_param)
+    filter = Admin::EditionFilter.new(Edition, build(:user), state: 'rejected', type: 'speech', author: user.to_param)
     assert_equal "John Doe's rejected speeches", filter.page_title
   end
 
   test "should generate page title when filtering by title" do
-    filter = EditionFilter.new(Edition, build(:user), title: 'test')
+    filter = Admin::EditionFilter.new(Edition, build(:user), title: 'test')
     assert_equal "Everyone's documents that match 'test'", filter.page_title
   end
 
   test "should generate page title when filtering by world location" do
     location = create(:world_location, name: 'Spain')
-    filter = EditionFilter.new(Edition, build(:user), world_location_ids: [location.to_param])
+    filter = Admin::EditionFilter.new(Edition, build(:user), world_location_ids: [location.to_param])
     assert_equal "Everyone's documents about Spain", filter.page_title
   end
 end
